# Conflicts:
#	src/main/java/com/jfshare/mvp/server/controller/AdminController.java
#	src/main/java/com/jfshare/mvp/server/model/TbJvjindouRule.java
#	src/main/java/com/jfshare/mvp/server/model/TbLevelInfo.java

+++ src/main/java/com/jfshare/mvp/server/mapper/TbLevelInfoMapper.java
@@ -1,12 +1,12 @@
 package com.jfshare.mvp.server.mapper;
 
-import com.jfshare.mvp.server.mapper.TbLevelInfo;
-import com.jfshare.mvp.server.mapper.TbLevelInfoExample;
+import com.jfshare.mvp.server.model.TbLevelInfo;
+import com.jfshare.mvp.server.model.TbLevelInfoExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbLevelInfoMapper {
-    long countByExample(TbLevelInfoExample example);
+    int countByExample(TbLevelInfoExample example);
 
     int deleteByExample(TbLevelInfoExample example);
