+++ src/main/java/com/jfshare/mvp/server/mapper/TbProductItemShowMapper.java
@@ -1,12 +1,12 @@
 package com.jfshare.mvp.server.mapper;
 
-import com.jfshare.mvp.server.mapper.TbProductItemShow;
-import com.jfshare.mvp.server.mapper.TbProductItemShowExample;
+import com.jfshare.mvp.server.model.TbProductItemShow;
+import com.jfshare.mvp.server.model.TbProductItemShowExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface TbProductItemShowMapper {
-    long countByExample(TbProductItemShowExample example);
+    int countByExample(TbProductItemShowExample example);
 
     int deleteByExample(TbProductItemShowExample example);
